+++ ChangeLog.md
@@ -10,6 +10,7 @@
 - Extend details extracted by WavPack handler
 - Fill database on startup
 - Filter resources by mimetype
+- Fix issue with stale objects in autoscan
 - Fix import and xml
 - Fix timed playlist import
 - Fix resources for containers